Media browser may now be called from anywhere by firing the `ha-media-player-browse-dialog` event
