--- src/components/media-player/show-media-browser-dialog.ts
@@ -1,27 +1,0 @@
-import { fireEvent } from "../../common/dom/fire_event";
-import {
-  MediaPickedEvent,
-  MediaPlayerBrowseAction,
-} from "../../data/media-player";
-
-export interface MediaPlayerBrowseDialogParams {
-  action: MediaPlayerBrowseAction;
-  entityId: string;
-  mediaPickedCallback: (pickedMedia: MediaPickedEvent) => void;
-  mediaContentId?: string;
-  mediaContentType?: string;
-}
-
-export const showMediaBrowserDialog = (
-  element: HTMLElement,
-  dialogParams: MediaPlayerBrowseDialogParams
-): void => {
-  fireEvent(element, "show-dialog", {
-    dialogTag: "dialog-media-player-browse",
-    dialogImport: () =>
-      import(
-        /* webpackChunkName: "dialog-media-player-browse" */ "./dialog-media-player-browse"
-      ),
-    dialogParams,
-  });
-};
